💄 style(topic): move view-toggle next to sort control in toolbar
